ui/search_manager: Move to ui namespace
search_manger is part of OZui now, so should live in src/ui
